+++ NeuralNetwork/NN.h
@@ -38,10 +38,10 @@ namespace NeuralNetwork {
         /**
          * Compute the Neural Network with given inputs
          *
-         * @param inputs_vector C double array of inputs
+         * @param inputs_vector C float array of inputs
          * @return a vector of weights
          */
-        double * compute(const double *inputs_vector);
+        float * compute(const float *inputs_vector);
 
         /**
          * Inits the Network from a topology
@@ -59,9 +59,9 @@ namespace NeuralNetwork {
     private:
         /**
          * Sets the inputs on the first layer
-         * @param inputs_vector Array of doubles to initiate the inputs
+         * @param inputs_vector Array of floats to initiate the inputs
          */
-        void set_inputs(const double *inputs_vector);
+        void set_inputs(const float *inputs_vector);
 
         /// Delete data
         void delete_layers();
